chore: add Prettier configuration file
A Prettier configuration file has been added to the project. This file defines the specifics of how Prettier should format the code, including settings for semi-colons, trailing commas, single quotes, tab width, and the use of tabs. It also specifies the use of two plugins: 'prettier-plugin-organize-imports' and 'prettier-plugin-jsdoc', for code organization and documentation respectively.