ci: add Build Pipeline workflow in GitHub Actions
This commit introduces a new GitHub Actions workflow named 'Build Pipeline'. This pipeline has various jobs for state verification, code building, testing, and generating a release, and it is triggered by push and pull request events, as well as manually. This will help automate and standardize procedures for testing and releasing.